+++ src/app/components/PageBasic/styled.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+export const MainContent = styled.div`
+  padding-top: 1em;
+`;
+
+export const SectionHead = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;