+++ src/app/layouts/AppLayout.tsx
@@ -61,7 +61,8 @@ const AppLayout: React.FunctionComponent<IAppLayout> = ({ children }) => {
   React.useEffect(() => {
     dispatch.setting.initSettingStore();
     dispatch.auth.checkLoginStatus();
-    const vSANMode = window.location.search.includes('vsan');
+    const vSANMode = history.location.search.includes('vsan');
+
     if (vSANMode) {
       dispatch.setting.setVSANMode();
     }