+++ src/app/pages/Inventory/Nodes/detail.styled.ts
@@ -15,3 +15,20 @@ export const LabelText = styled.span`
 export const TagContainer = styled.div`
   line-height: 2;
 `;
+
+export const EmptyDashboard = styled.div`
+  height: 300px;
+  width: 300px;
+  border: dashed 1px gray;
+  margin: 0 auto;
+  border-radius: 4px;
+  line-height: 300px;
+  text-align: center;
+  cursor: pointer;
+  color: gray;
+  font-size: 1.4rem;
+`;
+
+export const DashboardContainer = styled.div`
+  padding: 10px;
+`;